Restructure files for cli to allow reuse and fix import issues trying to import cli.cli
